security added to project

+++ src/main/kotlin/ir/balonet/support_service/service/MediatorService.kt
@@ -1,39 +1,77 @@
 package ir.balonet.support_service.service
 
+import ir.balonet.support_service.dao.AdminRepo
 import ir.balonet.support_service.dao.MediatorRepo
+import ir.balonet.support_service.exception.ForbiddenException
+import ir.balonet.support_service.exception.UnauthorizedException
 import ir.balonet.support_service.model.dto.MediatorDto
 import ir.balonet.support_service.model.entity.Mediator
+import ir.balonet.support_service.security.SecurityHelper
 import javassist.NotFoundException
+import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 
 
 @Service
-class MediatorService(val mediatorRepo: MediatorRepo) {
-    fun addMed(mediatorDto: MediatorDto): Mediator {
-        return mediatorRepo.save(Mediator.form(mediatorDto))
+class MediatorService(val mediatorRepo: MediatorRepo, val adminRepo: AdminRepo) {
+    fun addMed(token: String, mediatorDto: MediatorDto) {
+        if (adminRepo.existsByToken(token))
+            mediatorRepo.save(Mediator.form(mediatorDto))
+        else throw ForbiddenException(" don't have permission to access the requested")
     }
 
-    fun deleteMedById(medId: Long) {
-        if (mediatorRepo.existsById(medId))
-            mediatorRepo.deleteById(medId)
-        else throw NotFoundException("mediator by id:$medId not found")
+    fun deleteMedById(token: String, medId: Long) {
+        if (adminRepo.existsByToken(token)) {
+            if (mediatorRepo.existsById(medId))
+                mediatorRepo.deleteById(medId)
+            else throw NotFoundException("mediator by id:$medId not found")
+        } else throw ForbiddenException(" don't have permission to access the requested")
     }
 
-    fun updateMedById(medId: Long, mediatorDto: MediatorDto): Mediator {
-        if (mediatorRepo.existsById(medId)) {
-            val exMed = mediatorRepo.getById(medId).apply {
+    fun updateMedById(token: String,medId: Long, mediatorDto: MediatorDto): Mediator {
+        if (adminRepo.existsByToken(token)) {
+            val exMed = mediatorRepo.findByIdOrNull(medId)?:throw NotFoundException("mediator by id:$medId not found")
+            exMed.apply {
                 name = mediatorDto.name
                 password = mediatorDto.password
                 nationalId = mediatorDto.nationalId
             }
-            mediatorRepo.save(exMed)
-            return exMed
-        } else throw NotFoundException("mediator by id:$medId not found")
+            return mediatorRepo.save(exMed)
+
+        } else throw ForbiddenException(" don't have permission to access the requested")
+    }
+    fun updateMedByToken(token: String, mediatorDto: MediatorDto): Mediator {
+        if (mediatorRepo.existsByToken(token)) {
+            val exMed = mediatorRepo.findByToken(token).apply {
+                name = mediatorDto.name
+                password = mediatorDto.password
+                nationalId = mediatorDto.nationalId
+            }
+            return mediatorRepo.save(exMed)
+
+        } else throw ForbiddenException(" don't have permission to access the requested")
     }
 
-    fun getMedById(medId: Long): Mediator {
-        if (mediatorRepo.existsById(medId))
-            return mediatorRepo.getById(medId)
-        else throw NotFoundException("mediator by id:$medId not found")
+    fun getMedByToken(token: String): Mediator {
+        if (mediatorRepo.existsByToken(token)) {
+            return mediatorRepo.findByToken(token)
+        } else throw ForbiddenException(" don't have permission to access the requested")
+    }
+
+    fun getMedById(token: String, medId: Long): Mediator {
+        if (adminRepo.existsByToken(token)) {
+            if (mediatorRepo.existsById(medId))
+                return mediatorRepo.getById(medId)
+            else throw NotFoundException("mediator by id:$medId not found")
+        } else throw ForbiddenException(" don't have permission to access the requested")
+    }
+
+    fun login(nationalId: Long, password: String): String {
+        if (mediatorRepo.existsByNationalIdAndPassword(nationalId, password)) {
+            val generatedUser: Mediator = mediatorRepo.findByNationalIdAndPassword(nationalId, password).apply {
+                token = SecurityHelper.tokenGenerator()
+            }
+            return mediatorRepo.save(generatedUser).token
+        } else throw UnauthorizedException("wrong user or password")
     }
 }
