+++ src/main/kotlin/ir/balonet/support_service/service/UserService.kt
@@ -1,32 +1,42 @@
 package ir.balonet.support_service.service
 
+import ir.balonet.support_service.dao.AdminRepo
 import ir.balonet.support_service.dao.UserRepo
 import ir.balonet.support_service.exception.AlreadyExistException
+import ir.balonet.support_service.exception.ForbiddenException
 import ir.balonet.support_service.exception.NotFoundException
+import ir.balonet.support_service.exception.UnauthorizedException
 import ir.balonet.support_service.model.dto.UserDtoAdmin
 import ir.balonet.support_service.model.dto.UserDtoUser
 import ir.balonet.support_service.model.entity.User
+import ir.balonet.support_service.security.SecurityHelper
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 
 @Service
-class UserService(val userRepo: UserRepo) {
-    fun getAllUsers(): MutableList<User> {
-        return userRepo.findAll()
+class UserService(val userRepo: UserRepo, val adminRepo: AdminRepo) {
+    fun getAllUsers(token: String): MutableList<User> {
+        if (adminRepo.existsByToken(token))
+            return userRepo.findAll()
+        else throw ForbiddenException(" don't have permission to access the requested")
     }
 
-    fun getUserById(id: Long): User {
-        return userRepo.findByIdOrNull(id) ?: throw NotFoundException("there is no user by $id id")
+    fun getUserById(token: String, id: Long): User {
+        if (adminRepo.existsByToken(token)) {
+            return userRepo.findByIdOrNull(id) ?: throw NotFoundException("there is no user by $id id")
+        } else throw ForbiddenException(" don't have permission to access the requested")
     }
 
-    fun getByNameContaining(name: String): List<User> {
-        return userRepo.findByNameContaining(name)
+    fun getUserByToken(token: String): User {
+        if (userRepo.existsByToken(token)) {
+            return userRepo.findByToken(token)
+        } else throw ForbiddenException(" don't have permission to access the requested")
     }
 
-    fun getUserByNationalId(nationalId: Long): User {
-        if (userRepo.existsByNationalId(nationalId))
-            return getUserByNationalId(nationalId)
-        else throw NotFoundException("user by nationalId:$nationalId not found")
+    fun getByNameContaining(token: String, name: String): List<User> {
+        if (adminRepo.existsByToken(token))
+            return userRepo.findByNameContaining(name)
+        else throw ForbiddenException(" don't have permission to access the requested")
     }
 
     fun addUserByUser(newUser: UserDtoUser) {
@@ -35,46 +45,58 @@ class UserService(val userRepo: UserRepo) {
         userRepo.save(User.fromUser(newUser))
     }
 
-    fun addUserByAdmin(newUser: UserDtoAdmin) {
-        if (userRepo.existsByNationalId(newUser.nationalId))
-            throw AlreadyExistException(" this nationalId is registered before")
-        userRepo.save(User.fromAdmin(newUser))
-    }
-
-    fun updateByUser(updatedUser: UserDtoUser, id: Long): User {
-        //TODO get user id from tokenize user ind dont get in request
-        if (userRepo.existsById(id)) {
-            val exUser = userRepo.getById(id).apply {
-                name = updatedUser.name
-                password = updatedUser.password
-                nationalId = updatedUser.nationalId
-            }
-            userRepo.save(exUser)
-            return exUser
-        } else throw NotFoundException("there is no user by $id id")//
-
+    fun addUserByAdmin(token: String, newUser: UserDtoAdmin) {
+        if (adminRepo.existsByToken(token)) {
+            if (userRepo.existsByNationalId(newUser.nationalId))
+                throw AlreadyExistException(" this nationalId is registered before")
+            userRepo.save(User.fromAdmin(newUser))
+        }
 
     }
 
-    fun updateByAdmin(updatedUser: UserDtoAdmin, id: Long): User {
-        if (userRepo.existsById(id)) {
-            val exUser = userRepo.getById(id).apply {
+    fun updateByUser(token: String, updatedUser: UserDtoUser): User {
+        if (userRepo.existsByToken(token)) {
+            val exUser = userRepo.findByToken(token).apply {
                 name = updatedUser.name
                 password = updatedUser.password
                 nationalId = updatedUser.nationalId
-                isLocked = updatedUser.isLocked
             }
             return userRepo.save(exUser)
-        } else throw NotFoundException("there is no user by $id id")
+        } else throw ForbiddenException(" don't have permission to access the requested")
+    }
+
+    fun updateByAdmin(token: String, updatedUser: UserDtoAdmin, id: Long): User {
+        if (userRepo.existsByToken(token)) {
+            if (userRepo.existsById(id)) {
+                val exUser = userRepo.getById(id).apply {
+                    name = updatedUser.name
+                    password = updatedUser.password
+                    nationalId = updatedUser.nationalId
+                    isLocked = updatedUser.isLocked
+                }
+                return userRepo.save(exUser)
+            } else throw NotFoundException("there is no user by $id id")
+        } else throw ForbiddenException(" don't have permission to access the requested")
 
     }
 
-    fun deleteUserById(id: Long) {
-        if (userRepo.existsById(id))
-            userRepo.deleteById(id)
-        else
-            throw NotFoundException("there is no user by $id id")
+    fun deleteUserById(token: String, id: Long) {
+        if (userRepo.existsByToken(token)) {
+            if (userRepo.existsById(id))
+                userRepo.deleteById(id)
+            else
+                throw NotFoundException("there is no user by $id id")
+        } else throw ForbiddenException(" don't have permission to access the requested")
 
+    }
+
+    fun login(nationalId: Long, password: String): String {
+        if (userRepo.existsByNationalIdAndPassword(nationalId, password)) {
+            val generatedUser: User = userRepo.findByNationalIdAndPassword(nationalId, password).apply {
+                token = SecurityHelper.tokenGenerator()
+            }
+            return userRepo.save(generatedUser).token
+        } else throw UnauthorizedException("wrong user or password")
     }
 
 }
